+++ packages/epo-widget-lib/src/widgets/SupernovaThreeVector/Skymap/styles.ts
@@ -19,8 +19,14 @@ export const LatLabel = styled.text`
   font-size: 0.75em;
   text-anchor: end;
 `;
+
 export const LongLabel = styled.text`
   font-size: 0.75em;
   text-anchor: middle;
   dominant-baseline: central;
+`;
+
+export const ImageStackerWrapper = styled.foreignObject`
+  opacity: var(--image-stack-opacity, 0);
+  transition: opacity var(--DURATION_SLOW, 0.4s) ease;
 `;